backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg2-binary>=2.9,<3.0
pydantic-settings>=2.3,<3.0
passlib[bcrypt]>=1.7,<2.0
python-dotenv>=1.0,<2.0
email-validator>=2.1,<3.0
